v3 refactor: Copy()
