+++ marshal.go
@@ -96,23 +96,6 @@ type UsmSecurityParameters struct {
 	localAESSalt uint64
 }
 
-// Copy method for UsmSecurityParameters used to copy a SnmpV3SecurityParameters without knowing it's implementation
-func (sp *UsmSecurityParameters) Copy() SnmpV3SecurityParameters {
-	return &UsmSecurityParameters{AuthoritativeEngineID: sp.AuthoritativeEngineID,
-		AuthoritativeEngineBoots: sp.AuthoritativeEngineBoots,
-		AuthoritativeEngineTime:  sp.AuthoritativeEngineTime,
-		UserName:                 sp.UserName,
-		AuthenticationParameters: sp.AuthenticationParameters,
-		PrivacyParameters:        sp.PrivacyParameters,
-		AuthenticationProtocol:   sp.AuthenticationProtocol,
-		PrivacyProtocol:          sp.PrivacyProtocol,
-		AuthenticationPassphrase: sp.AuthenticationPassphrase,
-		PrivacyPassphrase:        sp.PrivacyPassphrase,
-		localDESSalt:             sp.localDESSalt,
-		localAESSalt:             sp.localAESSalt,
-	}
-}
-
 // SnmpPacket struct represents the entire SNMP Message or Sequence at the
 // application layer.
 type SnmpPacket struct {
